database: centralize row-representing struct with schema declaration
This makes it easier to make sure it's in sync when we do modifications
