+++ src/daemon/database/mod.rs
@@ -1,8 +1,6 @@
 pub mod actions;
-// We don't use all the interfaces just yet
-#[allow(dead_code)]
 pub mod interface;
-mod schema;
+pub mod schema;
 
 #[derive(PartialEq, Eq, Debug)]
 pub struct DatabaseError(pub String);
